requirements.txt
certifi==2018.11.29
chardet==3.0.4
cloudinary==1.15.0
defusedxml==0.5.0
Django==2.1.7
django-allauth==0.39.1
idna==2.8
mock==2.0.0
oauthlib==3.0.1
pbr==5.1.3
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
urllib3==1.24.1
